requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-JWT-Extended==4.5.2
Flask-Bcrypt==1.0.1
psycopg2-binary==2.9.7
python-binance==1.0.19
python-dotenv==1.0.0
requests==2.31.0
oauthlib==3.2.2
pyOpenSSL==23.2.0
cryptography==41.0.3
redis==5.0.0
celery==5.3.1
gunicorn==21.2.0
pytest==7.4.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.1.0
python-json-logger==2.0.7
